Removed `types` into folder and resolve some issues flagged
